move electron util to api
